requirements.txt
pandas
numpy
scipy
scikit-learn
yfinance
pyyaml
streamlit
plotly
pytest
